bsource and source - infix to postfix caculator using linkedlist stack
bsource is first way to making caculator using linkedlist stack.
Also it is not refactoring code.
It didn't have top struct.

But source.c is 2nd way to making caculator using linkedlist stack after refactoring of code.
It is more simple, efficient and easy to understading structure of linkedlist stack.

Also two source is okay to running.
This code is for data structure class assignment.